Change uses of int to size_t in MurmurHash

+++ src/support/MurmurHash3.c
@@ -58,11 +58,11 @@ FORCE_INLINE uint64_t fmix64 ( uint64_t k )
 
 //-----------------------------------------------------------------------------
 
-void MurmurHash3_x86_32 ( const void * key, int len,
+void MurmurHash3_x86_32 ( const void * key, size_t len,
                           uint32_t seed, void * out )
 {
   const uint8_t * data = (const uint8_t*)key;
-  const int nblocks = len / 4;
+  const size_t nblocks = len / 4;
 
   uint32_t h1 = seed;
 
@@ -74,7 +74,7 @@ void MurmurHash3_x86_32 ( const void * key, int len,
 
   const uint8_t * tail = data + nblocks*4;
 
-  for(int i = -nblocks; i; i++)
+  for(size_t i = -nblocks; i; i++)
   {
     uint32_t k1 = jl_load_unaligned_i32(tail + sizeof(uint32_t)*i);
 
@@ -112,11 +112,11 @@ void MurmurHash3_x86_32 ( const void * key, int len,
 
 //-----------------------------------------------------------------------------
 
-void MurmurHash3_x86_128 ( const void * key, const int len,
+void MurmurHash3_x86_128 ( const void * key, const size_t len,
                            uint32_t seed, void * out )
 {
   const uint8_t * data = (const uint8_t*)key;
-  const int nblocks = len / 16;
+  const size_t nblocks = len / 16;
 
   uint32_t h1 = seed;
   uint32_t h2 = seed;
@@ -133,7 +133,7 @@ void MurmurHash3_x86_128 ( const void * key, const int len,
 
   const uint8_t *tail = data + nblocks*16;
 
-  for(int i = -nblocks; i; i++)
+  for(size_t i = -nblocks; i; i++)
   {
     uint32_t k1 = jl_load_unaligned_i32(tail + sizeof(uint32_t)*(i*4 + 0));
     uint32_t k2 = jl_load_unaligned_i32(tail + sizeof(uint32_t)*(i*4 + 1));
@@ -218,11 +218,11 @@ void MurmurHash3_x86_128 ( const void * key, const int len,
 
 //-----------------------------------------------------------------------------
 
-void MurmurHash3_x64_128 ( const void * key, const int len,
+void MurmurHash3_x64_128 ( const void * key, const size_t len,
                            const uint32_t seed, void * out )
 {
   const uint8_t * data = (const uint8_t*)key;
-  const int nblocks = len / 16;
+  const size_t nblocks = len / 16;
 
   uint64_t h1 = seed;
   uint64_t h2 = seed;
@@ -233,7 +233,7 @@ void MurmurHash3_x64_128 ( const void * key, const int len,
   //----------
   // body
 
-  for(int i = 0; i < nblocks; i++)
+  for(size_t i = 0; i < nblocks; i++)
   {
     uint64_t k1 = jl_load_unaligned_i64(data + sizeof(uint64_t)*(i*2 + 0));
     uint64_t k2 = jl_load_unaligned_i64(data + sizeof(uint64_t)*(i*2 + 1));
